reorganizing core codes
